+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v1.1.0 (2026-04-17)
+
+- Added Lava Lamp effect with 8 bundled color themes.
+- Fixed Blur filter minimum distance changed to 2 pixels.
+
 ## v1.0.0 (2026-03-27)
 
 - Initial release with Rainbow Waves effect.